+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.0
 
+- Add `Lingua::Language[]` lookup by name, ISO 639-1 or ISO 639-3 code
 - Add `Lingua::Language.all`, `.names` and `.iso_codes` class methods
 
 ## 0.2.0